Fix plotting bug
Because the data class was using the first (i.e. earliest) entry to figure out the available information, some keys would be missing since they would only appear later. This was a particular problem with American Samoa, which didn't have a positive field and screwed up the plotting. This should fix the issue by getting the keys from the last entry.

Various other small fixes and tweaks as well.