Create percent positive graph
Modify the data class to calculate the percentage of new tests that are positive and then graph the average of that against the slope of the new cases vs current cases graph.

This creates a graph the measures success in two ways: is a state bending the curve and how well do we think we can rely on its testing.